pyproject.toml
[tool.poetry]
name = "myva"
version = "0.1.0"
description = ""
authors = ["Yvan J. Aquino <yaquino@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
flake8 = "^4.0.1"
google-cloud-secret-manager = "^2.11.1"
twilio = "^7.9.2"
googlemaps = "^4.6.0"
cxwebhooks = "^0.0.10"
pytest = "^7.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
